New hidden object field type in Gtk tables to hold the original string.
As strings with invalid characters are altered for display, a hidden field
is needed to hold the original value if the target string is used as a "key"
(i.e.: a value that is retrieved and used upon user action). Such a field
cannot be a string: this would cause Gtk to crash on conversion.
For this purpose, a new table field type is introduced to keep hidden objects.
S objects can be used to set such fields.

Modules that display "key" strings in Gtk tables now feature hidden
object fields to hold the real "key" value.