annotate: colorize row background according to author and revision.
This allows identifying more easily lines common to a single commit and
to the same author.

Background colors are randomly generated in the HSL color space to ease
luminance control (we generate light colors only), then converted to RGB
for gtk.

This commit also gets the file log in the loading action rather than in the
class initialization.